Tried using somebody's toobar control for tag tray and its really buggy...am just hiding the whole control for now. Added good optimization for tray query so tile inits are run as a batch. Added (hopefully) most logic for grid layout scrolling. That optimization helps grid layout a LOT since there's significantly more items. Need to adjust default vertical orientation size a bit so it holds items a little better in grid mode. Have a lot of the kinks worked out but should hopefully get to the rest tomorrow.

+++ MonkeyPaste.Avalonia/AttachedProperties/MpAvSizeObserverExtension.cs
@@ -95,10 +95,10 @@ namespace MonkeyPaste.Avalonia {
 
             void Control_EffectiveViewportChanged(object sender, global::Avalonia.Layout.EffectiveViewportChangedEventArgs? e) {
                 if (sender is Control control) {
-                    if(control.Bounds.Width != GetObservedWidth(control)) {
+                    if((int)control.Bounds.Width != (int)GetObservedWidth(control)) {
                         SetObservedWidth(control, control.Bounds.Width);
                     }
-                    if(control.Bounds.Height != GetObservedHeight(control)) {
+                    if((int)control.Bounds.Height != (int)GetObservedHeight(control)) {
                         SetObservedHeight(control, control.Bounds.Height);
                     }
